Conceptualized how I'm going to do the heap.
Basically using sized lists. memory can be reused as pointer.
Everything is going to be based off memorystack.

For memorystack... allocation should be threadsafe, but deallocation must be done in the order of allocation, so i don't really have to care if it's threadsafe.

For MemoryBlocks, it's a single allocation, so we don't really need to care about thread safety either. think of it as a normal raw array.

For the Heap, Will probably just use locks, for the returning.