Features/memory (#5)
* @added  Basic Structure of project and CMake file
@tofix  Linker for .lib file a bit wonky
@review Decide between auto adding files into projects vs manually adding files
[I support auto, only con is required to run CMake each time new files added]

* Added Bunch of placeholder memory stuff. Did a basic MemoryStack. We need a Logging System early.

* Memory Stack

* Adding some logging functions

* Added a bunch of stuff that i'll probably need to do some logs

* Basic MemoryManager class. Preallocated big stack. need to start reserving space for static data... string buffers. etc

* Circular String buffer. Basically... 1 giant buffer for all NonPersistentStrings. Used primarily for temp strings, like debug strings, strings to pass around. These strings should have a single frame life extent. If the strings get messed up, increase the size of the buffer.

* Added some logging functions

* Basic Logger... I think I'm gonna need some help rewriting the singletons.

* Conceptualized how I'm going to do the heap.

Basically using sized lists. memory can be reused as pointer.
Everything is going to be based off memorystack.

For memorystack... allocation should be threadsafe, but deallocation must be done in the order of allocation, so i don't really have to care if it's threadsafe.

For MemoryBlocks, it's a single allocation, so we don't really need to care about thread safety either. think of it as a normal raw array.

For the Heap, Will probably just use locks, for the returning.

* @removed MemoryPool
@implemented MemoryHeap
@todo MemoryHeap unit tests, Array Allocation
@feature Memory Alignment, Heap Allocation, Memory Tagging

* Added some new allocation functions. new, delete, allocate, deallocate, new has variaiadic .

added support for array allocate..

Yet to do Vector allocator.. not sure if i can figure it out actually.

I think i screwed up my git again. hopefully this works

* Hacked out basic test case thingy. will use this primitive method until we decide not to

* deleting code and putting it in a different place

* re-adding all the things you wrote

* <BREAKING BUILD>

Committing this so we can explore design decisions to be made while integrating
the memory management features into the uncmakified engine

* Converted memory/features no a nocmake version

Signed-off-by: Jonathan Chen <chen.enhan.jonathan@gmail.com>

* Fixed some mem tag stuff. I apparently forgot to pass the memtag string in...